more stuff.
better rotation buttons. and ui buttons and texbox for text support.
(backend not done).

As for the bug.  At around the coordinates (0, 6) sometimes the x
coordinate becomes a negative number due to the negative globe rotation.
This bug only is more frequent the higher negative globe rotation is.
I've locked the globe rotation for now so it doesn't go very high.